+++ src/main/java/com/ggang/be/api/group/registry/NearestGroupResponseStrategyRegistry.java
@@ -1,14 +1,15 @@
-package com.ggang.be.api.group.facade;
+package com.ggang.be.api.group.registry;
 
 import com.ggang.be.api.common.ResponseError;
 import com.ggang.be.api.exception.GongBaekException;
+import com.ggang.be.domain.group.vo.NearestGroup;
 import com.ggang.be.global.annotation.Registry;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 
 @Registry
 @RequiredArgsConstructor
-public class FindNearestGroupResponseStrategyRegistry {
+public class NearestGroupResponseStrategyRegistry {
 
     private final List<NearestGroupResponseStrategy> strategies;
