+++ src/main/java/com/ggang/be/api/userOnceGroup/service/UserOnceGroupService.java
@@ -1,6 +1,6 @@
 package com.ggang.be.api.userOnceGroup.service;
 
-import com.ggang.be.api.group.facade.NearestGroup;
+import com.ggang.be.domain.group.vo.NearestGroup;
 import com.ggang.be.domain.group.dto.ReadOnceGroupMember;
 import com.ggang.be.domain.group.onceGroup.OnceGroupEntity;
 import com.ggang.be.domain.group.onceGroup.dto.ReadOnceGroup;
