+++ src/main/java/com/ggang/be/domain/userEveryGroup/application/UserEveryGroupServiceImpl.java
@@ -3,7 +3,7 @@ package com.ggang.be.domain.userEveryGroup.application;
 
 import com.ggang.be.api.common.ResponseError;
 import com.ggang.be.api.exception.GongBaekException;
-import com.ggang.be.api.group.facade.NearestGroup;
+import com.ggang.be.domain.group.vo.NearestGroup;
 import com.ggang.be.api.userEveryGroup.service.UserEveryGroupService;
 import com.ggang.be.domain.constant.WeekDate;
 import com.ggang.be.domain.group.GroupVoMaker;
